views.py 파일 분리
1. views.py 파일 기능별로 분리
- 작성자,댓글,질문등
