--- board/views.py
@@ -1,239 +1,0 @@
-from django.shortcuts import render, get_object_or_404, redirect
-from django.http import HttpResponse
-from django.utils import timezone
-from .forms import QuestionForm, AnswerForm, CommentForm
-from django.core.paginator import Paginator
-from django.contrib.auth.decorators import login_required
-from django.contrib import messages
-from .models import Question, Answer, Comment
-
-def index(request):
-    page = request.GET.get('page', '1')  # 페이지
-    question_list = Question.objects.order_by('-create_date')
-    paginator = Paginator(question_list, 10)  # 페이지당 10개씩 보여주기
-    page_obj = paginator.get_page(page)
-
-    context = {'question_list': page_obj}
-    # ---------------------------------------------------------------------------- #
-    return render(request, 'board/question_list.html', context)
-
-
-def detail(request, question_id):
-    # 페이지가 없을 시 에러페이지 띄움
-    question = get_object_or_404(Question, pk=question_id)
-    context = {'question': question}
-    return render(request, 'board/question_detail.html', context)
-
-@login_required(login_url='common:login')
-def answer_create(request, question_id):
-    question = get_object_or_404(Question, pk=question_id)
-    if request.method == "POST":
-        form = AnswerForm(request.POST)
-        if form.is_valid():
-            answer = form.save(commit=False)
-            answer.author = request.user
-            answer.create_date = timezone.now()
-            answer.question = question
-            answer.save()
-            return redirect('board:detail', question_id=question.id)
-    else:
-        form = AnswerForm()
-    context = {'question': question, 'form': form}
-    return render(request, 'board/question_detail.html', context)
-
-@login_required(login_url='common:login')
-def question_create(request):
-    if request.method == 'POST':
-        form = QuestionForm(request.POST)
-        if form.is_valid():
-            question = form.save(commit=False)
-            question.author = request.user
-            question.create_date = timezone.now()
-            question.save()
-            return redirect('board:index')
-    else:
-        form = QuestionForm()
-    context = {'form': form}
-    return render(request, 'board/question_form.html', context)
-
-@login_required(login_url='common:login')
-def question_modify(request, question_id):
-    question = get_object_or_404(Question, pk=question_id)
-    if request.user != question.author:
-        messages.error(request, '수정권한이 없습니다')
-        return redirect('board:detail', question_id=question.id)
-
-    if request.method == "POST":
-        form = QuestionForm(request.POST, instance=question)
-        if form.is_valid():
-            question = form.save(commit=False)
-            question.author = request.user
-            question.modify_date = timezone.now()  # 수정일시 저장
-            question.save()
-            return redirect('board:detail', question_id=question.id)
-    else:
-        form = QuestionForm(instance=question)
-    context = {'form': form}
-    return render(request, 'board/question_form.html', context)
-
-@login_required(login_url='common:login')
-def question_delete(request, question_id):
-    question = get_object_or_404(Question, pk=question_id)
-    if request.user != question.author:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('board:detail', question_id=question.id)
-    question.delete()
-    return redirect('board:index')
-
-@login_required(login_url='common:login')
-def answer_modify(request, answer_id):
-    answer = get_object_or_404(Answer, pk=answer_id)
-    if request.user != answer.author:
-        messages.error(request, '수정권한이 없습니다')
-        return redirect('board:detail', question_id=answer.question.id)
-
-    if request.method == "POST":
-        form = AnswerForm(request.POST, instance=answer)
-        if form.is_valid():
-            answer = form.save(commit=False)
-            answer.author = request.user
-            answer.modify_date = timezone.now()
-            answer.save()
-            return redirect('board:detail', question_id=answer.question.id)
-    else:
-        form = AnswerForm(instance=answer)
-    context = {'answer': answer, 'form': form}
-    return render(request, 'board/answer_form.html', context)
-
-@login_required(login_url='common:login')
-def answer_delete(request, answer_id):
-    answer = get_object_or_404(Answer, pk=answer_id)
-    if request.user != answer.author:
-        messages.error(request, '삭제권한이 없습니다')
-    else:
-        answer.delete()
-    return redirect('board:detail', question_id=answer.question.id)
-
-@login_required(login_url='common:login')
-def comment_create_question(request, question_id):
-    question = get_object_or_404(Question, pk=question_id)
-    if request.method == "POST":
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.create_date = timezone.now()
-            comment.question = question
-            comment.save()
-            return redirect('board:detail', question_id=question.id)
-    else:
-        form = CommentForm()
-    context = {'form': form}
-    return render(request, 'board/comment_form.html', context)
-
-@login_required(login_url='common:login')
-def comment_modify_question(request, comment_id):
-    comment = get_object_or_404(Comment, pk=comment_id)
-    if request.user != comment.author:
-        messages.error(request, '댓글수정권한이 없습니다')
-        return redirect('board:detail', question_id=comment.question.id)
-
-    if request.method == "POST":
-        form = CommentForm(request.POST, instance=comment)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.modify_date = timezone.now()
-            comment.save()
-            return redirect('board:detail', question_id=comment.question.id)
-    else:
-        form = CommentForm(instance=comment)
-    context = {'form': form}
-    return render(request, 'board/comment_form.html', context)
-
-
-@login_required(login_url='common:login')
-def comment_delete_question(request, comment_id):
-    comment = get_object_or_404(Comment, pk=comment_id)
-    if request.user != comment.author:
-        messages.error(request, '댓글삭제권한이 없습니다')
-        return redirect('board:detail', question_id=comment.question.id)
-    else:
-        comment.delete()
-    return redirect('board:detail', question_id=comment.question.id)
-
-
-@login_required(login_url='common:login')
-def comment_create_answer(request, answer_id):
-    """
-    답글댓글등록
-    """
-    answer = get_object_or_404(Answer, pk=answer_id)
-    if request.method == "POST":
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.create_date = timezone.now()
-            comment.answer = answer
-            comment.save()
-            return redirect('board:detail', question_id=comment.answer.question.id)
-    else:
-        form = CommentForm()
-    context = {'form': form}
-    return render(request, 'board/comment_form.html', context)
-
-
-@login_required(login_url='common:login')
-def comment_modify_answer(request, comment_id):
-    """
-    답글댓글수정
-    """
-    comment = get_object_or_404(Comment, pk=comment_id)
-    if request.user != comment.author:
-        messages.error(request, '댓글수정권한이 없습니다')
-        return redirect('board:detail', question_id=comment.answer.question.id)
-
-    if request.method == "POST":
-        form = CommentForm(request.POST, instance=comment)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.modify_date = timezone.now()
-            comment.save()
-            return redirect('board:detail', question_id=comment.answer.question.id)
-    else:
-        form = CommentForm(instance=comment)
-    context = {'form': form}
-    return render(request, 'board/comment_form.html', context)
-
-
-@login_required(login_url='common:login')
-def comment_delete_answer(request, comment_id):
-    """
-    답글댓글삭제
-    """
-    comment = get_object_or_404(Comment, pk=comment_id)
-    if request.user != comment.author:
-        messages.error(request, '댓글삭제권한이 없습니다')
-        return redirect('board:detail', question_id=comment.answer.question.id)
-    else:
-        comment.delete()
-    return redirect('board:detail', question_id=comment.answer.question.id)
-
-# ------------------------------------------------------------------------
-
-# 제너릭 뷰 예시
-# class IndexView(generic.ListView):
-#     """
-#     목록 출력
-#     """
-#     def get_queryset(self):
-#         return Question.objects.order_by('-create_date')
-#
-#
-# class DetailView(generic.DetailView):
-#     """
-#     내용 출력
-#     """
-#     model = Question
