+++ board/views/base_views.py
@@ -1,0 +1,21 @@
+from django.core.paginator import Paginator
+from django.shortcuts import render, get_object_or_404
+
+from ..models import Question
+
+def index(request):
+    page = request.GET.get('page', '1')  # 페이지
+    question_list = Question.objects.order_by('-create_date')
+    paginator = Paginator(question_list, 10)  # 페이지당 10개씩 보여주기
+    page_obj = paginator.get_page(page)
+
+    context = {'question_list': page_obj}
+    # ---------------------------------------------------------------------------- #
+    return render(request, 'board/question_list.html', context)
+
+
+def detail(request, question_id):
+    # 페이지가 없을 시 에러페이지 띄움
+    question = get_object_or_404(Question, pk=question_id)
+    context = {'question': question}
+    return render(request, 'board/question_detail.html', context)
