+++ config/urls.py
@@ -15,11 +15,11 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from board import views
+from board.views import base_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('board/', include('board.urls')),
     path('common/', include('common.urls')),
-    path('', views.index, name='index'),
+    path('', base_views.index, name='index'),
 ]
